+++ src/data/tooltip-icons.js
@@ -1,14 +1,12 @@
 const searchButtonIcon = chrome.runtime.getURL('icons/button-icons/search.svg');
-
 const copyButtonIcon = chrome.runtime.getURL('icons/button-icons/copy.svg');
 const translateButtonIcon = chrome.runtime.getURL('icons/button-icons/translate.svg');
 const openLinkButtonIcon = chrome.runtime.getURL('icons/button-icons/open.svg');
-
 const cutButtonIcon = chrome.runtime.getURL('icons/button-icons/cut.svg');
 const pasteButtonIcon = chrome.runtime.getURL('icons/button-icons/paste.svg');
 const mapButtonIcon = chrome.runtime.getURL('icons/button-icons/map.svg');
 const emailButtonIcon = chrome.runtime.getURL('icons/button-icons/email.svg');
-const phoneIcon = chrome.runtime.getURL('icons/button-icons/phone.svg');
+const phoneIcon = chrome.runtime.getURL('icons/button-icons/call.svg');
 const boldTextIcon = chrome.runtime.getURL('icons/button-icons/bold.svg');
 const italicTextIcon = chrome.runtime.getURL('icons/button-icons/italic.svg');
 const clockIcon = chrome.runtime.getURL('icons/button-icons/time.svg');